Docker-Haproxy TLS (#38)
* Provisional proxy changes

Added ssl verify none to server_ops, also changed
http mode to change to tcp when SSL is configured.

Updated README.

* Rewording

* Removed unnessary mode change, and changes

After some investigation it was apparent that mode
`http` or `tcp` did not seem to make a difference with
haproxy. also discovered that `tls-ca-*` was the wrong
key to check for TLS. Changed to
`tls-cert-*` as an alternative.

* Docker TLS support.

Updated docs and checking of configuration keys.

* Review changes

Updated docs and changed TLS checks to check *-blob
instead of *-path.

* HAProxy changes

Added changes to config for verfied haproxy certificates.