Fix implementation of registry-port
Part of the charm code was assuming that the registry would always
run on port 5000 within the docker container, however the registry
configuration file was being rendered with the registry-port
so the host to container port mapping was broken.

Just always listen on 5000 and use the port mapping to implement
the registry-port option.