Allow enabling prometheus metrics
The docker registry can be configured to provide prometheus metrics, but
the charm currently does not have the option to enable this option. This
change adds the ability to enable this option, using the
`prometheus-metrics` and `debug-port` charm configuration options. There
is no way to only enable prometheus metrics, so enabling the
configuration option also enables other debug URLs such as /debug/vars.
It is expected that the other debugging URLs will be less relevant to
users of this charm, and so `prometheus-metrics` was chosen as the name
of the configuration option.

This change also adds a small bit of fanciness by updating the ports
displayed in the output of `juju status` dynamically based on their
values and the value of the `prometheus-metrics` option.

Fixes https://bugs.launchpad.net/layer-docker-registry/+bug/2034475